+++ test/e2e/app-websocket.ts
@@ -6,7 +6,7 @@ import {
   AppWebsocket,
   CallZomeRequest,
   CellType,
-  CloneId,
+  CloneIdHelper,
   CreateCloneCellRequest,
   fakeAgentPubKey,
   RoleName,
@@ -222,7 +222,7 @@ test(
     const cloneCell = await appWs.createCloneCell(createCloneCellParams);
     await admin.authorizeSigningCredentials(cloneCell.cell_id);
 
-    const expectedCloneId = new CloneId(ROLE_NAME, 0).toString();
+    const expectedCloneId = new CloneIdHelper(ROLE_NAME, 0).toString();
     t.equal(cloneCell.clone_id, expectedCloneId, "correct clone id");
     assert(info.cell_info[ROLE_NAME][0].type === CellType.Provisioned);
     t.deepEqual(
@@ -286,30 +286,6 @@ test(
     });
     await appWs.callZome(params);
     t.pass("re-enabled clone can be called");
-  })
-);
-
-test(
-  "can fetch network info",
-  withConductor(ADMIN_PORT, async (t) => {
-    const { client: appWs, cell_id } = await createAppWsAndInstallApp(
-      ADMIN_PORT
-    );
-
-    const response = await appWs.networkInfo({
-      dnas: [cell_id[0]],
-    });
-
-    t.deepEqual(response, [
-      {
-        fetch_pool_info: { op_bytes_to_fetch: 0, num_ops_to_fetch: 0 },
-        current_number_of_peers: 1,
-        arc_size: 1,
-        total_network_peers: 1,
-        bytes_since_last_time_queried: 1838,
-        completed_rounds_since_last_time_queried: 0,
-      },
-    ]);
   })
 );