+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## \[Unreleased\]
 
 ### Added
+- Call `DumpNetworkStats` was added to Admin + App websockets.
+- Call `DumpNetworkMetrics` was added to Admin + App websockets.
 ### Fixed
 ### Changed
 ### Removed
+- Call `NetworkInfo` was removed from the Conductor API.
+- Call `GetCompatibleCells` was moved to a feature.
 
 ## 2025-03-11: v0.19.0-dev.8